refactor(balancing): deepen template and model modules
