+++ libs/gda-balancing/src/gda_balancing/application/model_inspect.py
@@ -4,8 +4,10 @@ from dataclasses import dataclass
 from typing import Any
 
 from gda_balancing.domain.artifact_set import ArtifactSetMemberSpec
-from gda_balancing.domain.model.inspection import read_model_explanation
-from gda_balancing.domain.model.inspection_types import ModelInspectAdmissionError
+from gda_balancing.domain.model import (
+    ModelInspectAdmissionError,
+    read_model_explanation,
+)
 from gda_balancing.domain.diagnostics import Schema2RefusalReport, ingress_refusal
