+++ libs/gda-balancing/src/gda_balancing/domain/model/_resolution.py
