+++ libs/gda-balancing/src/gda_balancing/domain/template/__init__.py
@@ -1,0 +1,19 @@
+"""Template release admission, validation, and packaged Adapters."""
+
+from .core import (
+    TemplateProvider,
+    load_admitted_template,
+    prepare_template_instantiation,
+    template_refusal,
+    validate_template_release,
+)
+from .quantity_minimal import minimal_release
+
+__all__ = (
+    "TemplateProvider",
+    "load_admitted_template",
+    "minimal_release",
+    "prepare_template_instantiation",
+    "template_refusal",
+    "validate_template_release",
+)
