+++ libs/gda-balancing/src/gda_balancing/interfaces/cli/model_build.py
@@ -13,7 +13,7 @@ from gda_balancing.interfaces.cli.artifact_set import ArtifactSetMemberLocator
 from gda_balancing.interfaces.cli.model_fixtures import VALID_MODEL_SOURCE
 from gda_balancing.interfaces.cli.path_contracts import reject_input_aliasing
 from gda_balancing.domain.diagnostics import Schema2RefusalReport
-from gda_balancing.domain.model.resolution import MODEL_REFUSAL_CATALOG
+from gda_balancing.domain.model import MODEL_REFUSAL_CATALOG
 from gda_balancing.interfaces.cli.surface import descriptor_identity
