+++ libs/gda-balancing/tests/test_schema2_template_cli.py
@@ -18,7 +18,6 @@ from gda_balancing.interfaces.cli.template_instantiation import (
     template_instantiate_handler,
 )
 from gda_balancing.domain.template import (
-    _member_schema_identities,
     minimal_release,
     validate_template_release,
 )
@@ -34,9 +33,11 @@ from gda_balancing.domain.authority.context import (
 from gda_balancing.domain.authority.graph import derive_language_index
 from gda_balancing.domain.canonical import JsonValue, canonical_bytes, content_identity
 from gda_balancing.domain.diagnostics import Schema2RefusalReport
-from gda_balancing.domain.model.resolution import CheckedModel
-from gda_balancing.domain.model.lowering import checked_model_template_facts
-from gda_balancing.domain.model.checking import check_model_source_value
+from gda_balancing.domain.model import (
+    CheckedModel,
+    check_model_source_value,
+    checked_model_template_facts,
+)
 from gda_balancing.domain.authority.projections import wire_schema_projection
 from gda_balancing.domain.authority.package_semantics import (
     package_runtime_semantic_closure,
@@ -44,6 +45,7 @@ from gda_balancing.domain.authority.package_semantics import (
 from gda_balancing.domain.wire_schema import (
     artifact_wire_schema_identity,
     wire_schema_identity as schema_definition_identity,
+    wire_schema_identity_for_kind,
 )
 
 
@@ -931,7 +933,7 @@ def test_template_schema_identity_refuses_a_missing_authority_contract():
         ValueError,
         match=f"exact wire-schema identity domain is unavailable for {schema_kind}",
     ):
-        _member_schema_identities(language_bundle)
+        wire_schema_identity_for_kind(language_bundle, schema_kind)
 
 
 def test_every_wire_schema_consumer_projects_an_extension_owned_identity_domain():
@@ -960,12 +962,6 @@ def test_every_wire_schema_consumer_projects_an_extension_owned_identity_domain(
         if row["artifact_kind"] == "model-source-package"
     )
 
-    assert _member_schema_identities(context.language_bundle)[
-        "model-source-package"
-    ] == artifacts_module.wire_schema_identity(
-        context.language_bundle,
-        "model-source-package",
-    )
     assert (
         artifacts_module.wire_schema_identity(
             context.language_bundle,
@@ -1126,7 +1122,10 @@ def test_template_admits_a_member_whose_artifact_and_schema_kinds_differ():
 def test_template_refuses_an_artifact_kind_that_shadows_a_standalone_schema():
     authority = authority_set()
     language_bundle = authority["language_bundle"]
-    original_identity = _member_schema_identities(language_bundle)["negative-vector"]
+    original_identity = wire_schema_identity_for_kind(
+        language_bundle,
+        "negative-vector",
+    )
     _install_negative_vector_artifact_contract(
         language_bundle,
         retain_standalone=True,
@@ -1136,12 +1135,7 @@ def test_template_refuses_an_artifact_kind_that_shadows_a_standalone_schema():
         ValueError,
         match="wire-schema kind authority is ambiguous: negative-vector",
     ):
-        _member_schema_identities(language_bundle)
-    with pytest.raises(
-        ValueError,
-        match="wire-schema kind authority is ambiguous: negative-vector",
-    ):
-        artifact_wire_schema_identity(language_bundle, "negative-vector")
+        wire_schema_identity_for_kind(language_bundle, "negative-vector")
 
     assert original_identity == content_identity(
         "negative-vector-wire-schema-v2",
